Fix colorized output in workflow summary
Remove colorized output from the workflow summary. This means only
printing the summary line without color and calling `setFailed` with
this line. The actual issues are only printed in the job output now, so
users will need to click through to it.

This isn't what I would have liked. GitHub should instead remove the
colorized output formatting from the summary page if they don't support
it there, or provide an option to do so.